Clean up chat service and add env example
