+++ src/chat/env.example
@@ -1,0 +1,6 @@
+HOST=0.0.0.0
+PORT=3006
+
+DATABASE_URL=file:./prisma/dev.db
+
+SERVICE_NAME=chat
