+++ src/chat/src/index.ts
@@ -1,78 +1,25 @@
 import fastify from 'fastify';
-import cors from '@fastify/cors';
-import helmet from '@fastify/helmet';
 import { getPrismaClient, disconnectPrisma } from './db/prisma.client.js';
 import { loggerOptions } from './config/logger.js';
-import { ChannelDao } from './dao/channel.dao.js';
-import { MessageDao } from './dao/message.dao.js';
-import { NotificationService } from './services/notification.js';
-import { BlockService } from './services/block.js';
-import { MatchmakingClient } from './services/matchmaking-client.js';
-import { UserClient } from './services/user-client.js';
-import { ChatService } from './services/chat.js';
+import { createClients } from './config/clients.js';
+import { createServices } from './config/services.js';
+import { registerHealthRoutes } from './routes/health.js';
 import { registerChannelRoutes } from './routes/channels.js';
 import { registerMessageRoutes } from './routes/messages.js';
 import { registerInternalRoutes } from './routes/internal.js';
 
 const server = fastify({ logger: loggerOptions });
-
-// Register plugins
-await server.register(cors, {
-  origin: process.env.ALLOWED_ORIGINS?.split(',') || ['http://localhost:3000'],
-  credentials: true
-});
-
-await server.register(helmet, {
-  contentSecurityPolicy: false
-});
-
-// Initialize services
 const prisma = getPrismaClient();
-const channelDao = new ChannelDao(prisma);
-const messageDao = new MessageDao(prisma);
-const notificationService = new NotificationService(channelDao, server.log);
-const blockService = new BlockService(server.log);
-const matchmakingClient = new MatchmakingClient(
-  process.env.MATCHMAKING_URL || 'http://localhost:3005'
-);
-const userClient = new UserClient(server.log);
-const chatService = new ChatService(channelDao, messageDao, notificationService, blockService, server.log, matchmakingClient, userClient);
+const clients = createClients(server.log);
+const services = createServices(prisma, clients, server.log);
 
-// Health checks
-server.get('/health', async () => ({
-  status: 'healthy',
-  service: 'chat-service',
-  timestamp: new Date().toISOString(),
-  uptime: process.uptime()
-}));
+await registerHealthRoutes(server, prisma);
+await registerChannelRoutes(server, services.channelService);
+await registerMessageRoutes(server, services.messageService, services.matchAckService);
+await registerInternalRoutes(server, services.matchAckService, services.channelService, services.messageService);
 
-server.get('/health/detailed', async () => {
-  let dbHealthy = false;
-  try {
-    await prisma.$queryRaw`SELECT 1`;
-    dbHealthy = true;
-  } catch (error) {
-    server.log.error({ error }, 'Database health check failed');
-  }
-
-  return {
-    status: dbHealthy ? 'healthy' : 'degraded',
-    service: 'chat-service',
-    timestamp: new Date().toISOString(),
-    uptime: process.uptime(),
-    database: dbHealthy ? 'connected' : 'disconnected'
-  };
-});
-
-// Register routes
-await registerChannelRoutes(server, chatService);
-await registerMessageRoutes(server, chatService);
-await registerInternalRoutes(server, chatService);
-
-// Graceful shutdown
 const shutdown = async (signal: string) => {
   server.log.info(`Received ${signal}, shutting down gracefully`);
-
   try {
     await disconnectPrisma();
     await server.close();
@@ -87,12 +34,10 @@ const shutdown = async (signal: string) => {
 process.on('SIGTERM', () => shutdown('SIGTERM'));
 process.on('SIGINT', () => shutdown('SIGINT'));
 
-// Start server
 const start = async () => {
   try {
     const host = process.env.HOST || '0.0.0.0';
     const port = parseInt(process.env.PORT || '3006', 10);
-
     await server.listen({ port, host });
     server.log.info(`Chat service listening on ${host}:${port}`);
   } catch (error) {
